Deserialization of Status struct.
There is a clean and beautifull way of doing this but at this quick hack
together should atleast send correct results via HTTP even though the
strings in the struct contain extra \x00's. This will be fixed later.

The tests will still fail because of this but I can atleast run this
on my own minecraft server already. Have to read trough the server docs
some more and the Go package docs to implement a fully correct version.